+++ server/storage_index_test.go
@@ -55,7 +55,7 @@ func TestLocalStorageIndex_Write(t *testing.T) {
 	})
 	valueThree := string(valueThreeBytes)
 
-	storageIdx, err := NewLocalStorageIndex(logger, db, &StorageConfig{})
+	storageIdx, err := NewLocalStorageIndex(logger, db, &StorageConfig{}, metrics)
 	if err != nil {
 		t.Fatal(err.Error())
 	}
@@ -207,7 +207,7 @@ func TestLocalStorageIndex_Write(t *testing.T) {
 			UpdateTime:      timestamppb.New(ts),
 		}
 
-		storageIdx, err := NewLocalStorageIndex(logger, db, &StorageConfig{})
+		storageIdx, err := NewLocalStorageIndex(logger, db, &StorageConfig{}, metrics)
 		if err != nil {
 			t.Fatal(err.Error())
 		}
@@ -331,7 +331,7 @@ func TestLocalStorageIndex_List(t *testing.T) {
 		})
 		valueThree := string(valueThreeBytes)
 
-		storageIdx, err := NewLocalStorageIndex(logger, db, &StorageConfig{})
+		storageIdx, err := NewLocalStorageIndex(logger, db, &StorageConfig{}, metrics)
 		if err != nil {
 			t.Fatal(err.Error())
 		}
@@ -424,7 +424,7 @@ func TestLocalStorageIndex_List(t *testing.T) {
 		})
 		valueThree := string(valueThreeBytes)
 
-		storageIdx, err := NewLocalStorageIndex(logger, db, &StorageConfig{})
+		storageIdx, err := NewLocalStorageIndex(logger, db, &StorageConfig{}, metrics)
 		if err != nil {
 			t.Fatal(err.Error())
 		}
@@ -508,7 +508,7 @@ func TestLocalStorageIndex_Delete(t *testing.T) {
 	})
 	valueOne := string(valueOneBytes)
 
-	storageIdx, err := NewLocalStorageIndex(logger, db, &StorageConfig{})
+	storageIdx, err := NewLocalStorageIndex(logger, db, &StorageConfig{}, metrics)
 	if err != nil {
 		t.Fatal(err.Error())
 	}